+++ saleor/graphql/product/tests/test_product.py
@@ -27,7 +27,6 @@ from ....attribute.models import Attribute, AttributeValue
 from ....attribute.utils import associate_attribute_values_to_instance
 from ....core.taxes import TaxType
 from ....core.units import MeasurementUnits, WeightUnits
-from ....core.utils.editorjs import clean_editor_js
 from ....order import OrderEvents, OrderStatus
 from ....order.models import OrderEvent, OrderLine
 from ....plugins.manager import PluginsManager, get_plugins_manager
@@ -44,7 +43,7 @@ from ....product.models import (
     ProductVariant,
     ProductVariantChannelListing,
 )
-from ....product.search import prepare_product_search_document_value
+from ....product.search import prepare_product_search_vector_value
 from ....product.tasks import update_variants_names
 from ....product.tests.utils import create_image, create_pdf_file_with_image_ext
 from ....product.utils.availability import get_variant_availability
@@ -2350,8 +2349,8 @@ def test_products_query_with_filter_category_and_search(
     product.save()
 
     for pr in [product, second_product]:
-        pr.search_document = prepare_product_search_document_value(pr)
-    Product.objects.bulk_update([product, second_product], ["search_document"])
+        pr.search_vector = prepare_product_search_vector_value(pr)
+    Product.objects.bulk_update([product, second_product], ["search_vector"])
 
     category_id = graphene.Node.to_global_id("Category", category.id)
     variables = {"filter": {"categories": [category_id], "search": product.name}}
@@ -2482,10 +2481,8 @@ def test_products_query_with_filter(
         channel=channel_USD,
         is_published=False,
     )
-    second_product.search_document = prepare_product_search_document_value(
-        second_product
-    )
-    second_product.save(update_fields=["search_document"])
+    second_product.search_vector = prepare_product_search_vector_value(second_product)
+    second_product.save(update_fields=["search_vector"])
     variables = {"filter": products_filter, "channel": channel_USD.slug}
     staff_api_client.user.user_permissions.add(permission_manage_products)
     response = staff_api_client.post_graphql(query_products_with_filter, variables)
@@ -2595,10 +2592,10 @@ def test_products_query_with_filter_search_by_dropdown_attribute_value(
 
     product_with_dropdown_attr.refresh_from_db()
 
-    product_with_dropdown_attr.search_document = prepare_product_search_document_value(
+    product_with_dropdown_attr.search_vector = prepare_product_search_vector_value(
         product_with_dropdown_attr
     )
-    product_with_dropdown_attr.save(update_fields=["search_document"])
+    product_with_dropdown_attr.save(update_fields=["search_document", "search_vector"])
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
@@ -2618,7 +2615,7 @@ def test_products_query_with_filter_search_by_dropdown_attribute_value(
 
 
 @pytest.mark.parametrize(
-    "search_value", ["eco mode", "ECO Performance", "performa", "mod"]
+    "search_value", ["eco mode", "ECO Performance", "performant*", "modes"]
 )
 def test_products_query_with_filter_search_by_multiselect_attribute_value(
     search_value,
@@ -2657,10 +2654,10 @@ def test_products_query_with_filter_search_by_multiselect_attribute_value(
 
     product_with_multiselect_attr.refresh_from_db()
 
-    product_with_multiselect_attr.search_document = (
-        prepare_product_search_document_value(product_with_multiselect_attr)
+    product_with_multiselect_attr.search_vector = prepare_product_search_vector_value(
+        product_with_multiselect_attr
     )
-    product_with_multiselect_attr.save(update_fields=["search_document"])
+    product_with_multiselect_attr.save(update_fields=["search_vector"])
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
@@ -2705,10 +2702,10 @@ def test_products_query_with_filter_search_by_rich_text_attribute(
 
     product_with_rich_text_attr.refresh_from_db()
 
-    product_with_rich_text_attr.search_document = prepare_product_search_document_value(
+    product_with_rich_text_attr.search_vector = prepare_product_search_vector_value(
         product_with_rich_text_attr
     )
-    product_with_rich_text_attr.save(update_fields=["search_document"])
+    product_with_rich_text_attr.save(update_fields=["search_vector"])
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
@@ -2756,10 +2753,10 @@ def test_products_query_with_filter_search_by_numeric_attribute_value(
 
     product_with_numeric_attr.refresh_from_db()
 
-    product_with_numeric_attr.search_document = prepare_product_search_document_value(
+    product_with_numeric_attr.search_vector = prepare_product_search_vector_value(
         product_with_numeric_attr
     )
-    product_with_numeric_attr.save(update_fields=["search_document"])
+    product_with_numeric_attr.save(update_fields=["search_vector"])
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
@@ -2803,10 +2800,10 @@ def test_products_query_with_filter_search_by_numeric_attribute_value_without_un
 
     product_with_numeric_attr.refresh_from_db()
 
-    product_with_numeric_attr.search_document = prepare_product_search_document_value(
+    product_with_numeric_attr.search_vector = prepare_product_search_vector_value(
         product_with_numeric_attr
     )
-    product_with_numeric_attr.save(update_fields=["search_document"])
+    product_with_numeric_attr.save(update_fields=["search_vector"])
 
     variables = {"filter": {"search": "13456"}, "channel": channel_USD.slug}
 
@@ -2825,7 +2822,7 @@ def test_products_query_with_filter_search_by_numeric_attribute_value_without_un
     assert products[0]["node"]["name"] == product_with_numeric_attr.name
 
 
-@pytest.mark.parametrize("search_value", ["2020", "2020 10 10", "2020-10-10"])
+@pytest.mark.parametrize("search_value", ["2020", "2020-10-10"])
 def test_products_query_with_filter_search_by_date_attribute_value(
     search_value,
     query_products_with_filter,
@@ -2851,10 +2848,10 @@ def test_products_query_with_filter_search_by_date_attribute_value(
 
     product_with_date_attr.refresh_from_db()
 
-    product_with_date_attr.search_document = prepare_product_search_document_value(
+    product_with_date_attr.search_vector = prepare_product_search_vector_value(
         product_with_date_attr
     )
-    product_with_date_attr.save(update_fields=["search_document"])
+    product_with_date_attr.save(update_fields=["search_vector"])
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
@@ -2873,7 +2870,7 @@ def test_products_query_with_filter_search_by_date_attribute_value(
     assert products[0]["node"]["name"] == product_with_date_attr.name
 
 
-@pytest.mark.parametrize("search_value", ["2020", "2020 10 10", "2020-10-10", "22:20"])
+@pytest.mark.parametrize("search_value", ["2020", "2020-10-10", "22:20"])
 def test_products_query_with_filter_search_by_date_time_attribute_value(
     search_value,
     query_products_with_filter,
@@ -2899,10 +2896,10 @@ def test_products_query_with_filter_search_by_date_time_attribute_value(
 
     product_with_date_time_attr.refresh_from_db()
 
-    product_with_date_time_attr.search_document = prepare_product_search_document_value(
+    product_with_date_time_attr.search_vector = prepare_product_search_vector_value(
         product_with_date_time_attr
     )
-    product_with_date_time_attr.save(update_fields=["search_document"])
+    product_with_date_time_attr.save(update_fields=["search_vector"])
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
@@ -4025,10 +4022,6 @@ def test_create_product(
 
     product = Product.objects.first()
     created_webhook_mock.assert_called_once_with(product)
-    assert product.search_document
-    assert product_name.lower() in product.search_document
-    assert color_value_slug.lower() in product.search_document
-    assert non_existent_attr_value.lower() in product.search_document
     updated_webhook_mock.assert_not_called()
 
 
@@ -4079,7 +4072,6 @@ def test_create_product_description_plaintext(
 
     product = Product.objects.all().first()
     assert product.description_plaintext == description
-    assert description in product.search_document
 
 
 def test_create_product_with_rich_text_attribute(
@@ -4156,13 +4148,6 @@ def test_create_product_with_rich_text_attribute(
 
     for attr_data in data["product"]["attributes"]:
         assert attr_data in expected_attributes_data
-    product = Product.objects.first()
-    assert product.search_document
-    assert product_name.lower() in product.search_document
-    assert (
-        clean_editor_js(rich_text_value, to_string=True).lower()
-        in product.search_document
-    )
 
 
 def test_create_product_no_value_for_rich_text_attribute(
@@ -4273,10 +4258,6 @@ def test_create_product_with_date_time_attribute(
     }
 
     assert expected_attributes_data in data["product"]["attributes"]
-    product = Product.objects.first()
-    assert product.search_document
-    assert product_name.lower() in product.search_document
-    assert str(value.isoformat()).lower() in product.search_document
 
 
 @freeze_time(datetime(2020, 5, 5, 5, 5, 5, tzinfo=pytz.utc))
@@ -4338,10 +4319,6 @@ def test_create_product_with_date_attribute(
     }
 
     assert expected_attributes_data in data["product"]["attributes"]
-    product = Product.objects.first()
-    assert product.search_document
-    assert product_name.lower() in product.search_document
-    assert str(value).lower() in product.search_document
 
 
 def test_create_product_no_value_for_date_attribute(
@@ -4547,10 +4524,11 @@ def test_search_product_by_description_and_name(
 
     product_list.append(product)
     for prod in product_list:
-        prod.search_document = prepare_product_search_document_value(prod)
+        prod.search_vector = prepare_product_search_vector_value(prod)
 
     Product.objects.bulk_update(
-        product_list, ["search_document", "name", "description_plaintext"]
+        product_list,
+        ["search_document", "search_vector", "name", "description_plaintext"],
     )
 
     variables = {
@@ -4572,22 +4550,23 @@ def test_search_product_by_description_and_name(
 
 
 def test_sort_product_by_rank_without_search(
-    user_api_client, product_list, product, channel_USD, category, product_type
+    user_api_client, product_list, channel_USD
 ):
-    product_count = Product.objects.count()
-
     variables = {
         "sortBy": {"field": "RANK", "direction": "DESC"},
         "channel": channel_USD.slug,
     }
     response = user_api_client.post_graphql(SEARCH_PRODUCTS_QUERY, variables)
-    content = get_graphql_content(response)
-    data = content["data"]["products"]["edges"]
-    assert len(data) == product_count
+    content = get_graphql_content(response, ignore_errors=True)
+    assert "errors" in content
+    assert (
+        content["errors"][0]["message"]
+        == "Sorting by Rank is available only with searching."
+    )
 
 
 def test_search_product_by_description_and_name_without_sort_by(
-    user_api_client, product_list, product, channel_USD, category, product_type
+    user_api_client, product_list, product, channel_USD
 ):
     product.description_plaintext = "new big new product"
 
@@ -4598,10 +4577,11 @@ def test_search_product_by_description_and_name_without_sort_by(
 
     product_list.append(product)
     for prod in product_list:
-        prod.search_document = prepare_product_search_document_value(prod)
+        prod.search_vector = prepare_product_search_vector_value(prod)
 
     Product.objects.bulk_update(
-        product_list, ["search_document", "name", "description_plaintext"]
+        product_list,
+        ["search_vector", "name", "description_plaintext"],
     )
 
     variables = {
@@ -4634,10 +4614,11 @@ def test_search_product_by_description_and_name_and_use_cursor(
 
     product_list.append(product)
     for prod in product_list:
-        prod.search_document = prepare_product_search_document_value(prod)
+        prod.search_vector = prepare_product_search_vector_value(prod)
 
     Product.objects.bulk_update(
-        product_list, ["search_document", "name", "description_plaintext"]
+        product_list,
+        ["search_vector", "name", "description_plaintext"],
     )
 
     variables = {
@@ -4662,7 +4643,6 @@ def test_search_product_by_description_and_name_and_use_cursor(
     data = content["data"]["products"]["edges"]
 
     assert len(data) == 2
-    assert {node["node"]["name"] for node in data} == {product_1.name, product_2.name}
 
 
 @freeze_time("2020-03-18 12:00:00")
@@ -6373,10 +6353,6 @@ def test_update_product(
 
     updated_webhook_mock.assert_called_once_with(product)
     created_webhook_mock.assert_not_called()
-    product.refresh_from_db()
-    assert product.search_document
-    assert attr_value.lower() in product.search_document
-    assert product_name.lower() in product.search_document
 
 
 def test_update_and_search_product_by_description(
@@ -6687,10 +6663,6 @@ def test_update_product_with_numeric_attribute_value(
     assert expected_att_data in attributes
 
     updated_webhook_mock.assert_called_once_with(product)
-
-    product.refresh_from_db()
-    assert product.search_document
-    assert f"{new_value}{numeric_attribute.unit}" in product.search_document
 
 
 @patch("saleor.plugins.manager.PluginsManager.product_updated")
@@ -9189,9 +9161,6 @@ def test_product_type_update_mutation(
     require_shipping = False
     product_type_id = graphene.Node.to_global_id("ProductType", product_type.id)
 
-    product_attr = product.attributes.first()
-    value = product_attr.values.first()
-
     # Test scenario: remove all product attributes using [] as input
     # but do not change variant attributes
     product_attributes = []
@@ -9218,10 +9187,6 @@ def test_product_type_update_mutation(
     assert data["isShippingRequired"] == require_shipping
     assert not data["productAttributes"]
     assert len(data["variantAttributes"]) == (variant_attributes.count())
-
-    product.refresh_from_db()
-    assert product.search_document
-    assert value.name not in product.search_document
 
 
 def test_product_type_update_mutation_not_valid_attributes(
@@ -9301,9 +9266,9 @@ UPDATE_PRODUCT_TYPE_SLUG_MUTATION = """
         (None, "", "Slug value cannot be blank."),
     ],
 )
-@patch("saleor.product.search.update_products_search_document")
+@patch("saleor.product.search.update_products_search_vector")
 def test_update_product_type_slug(
-    update_products_search_document_mock,
+    update_products_search_vector_mock,
     staff_api_client,
     product_type,
     permission_manage_product_types_and_attributes,
@@ -9327,7 +9292,7 @@ def test_update_product_type_slug(
     if not error_message:
         assert not errors
         assert data["productType"]["slug"] == expected_slug
-        update_products_search_document_mock.assert_not_called()
+        update_products_search_vector_mock.assert_not_called()
     else:
         assert errors
         assert errors[0]["field"] == "slug"