+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix access to own resources by App - #9425 by @korycins
 - Introduce custom prices - #9393 by @IKarbowiak
   - Add `HANDLE_CHECKOUTS` permission (only for apps)
+- Use full-text search for products search API - #9344 by @patrys
 
 
 # 3.1.7